+++ packages/bitcore-hardware/src/base.ts
@@ -1,8 +1,8 @@
-import { BaseParams, SignParams } from './types/paramTypes.js';
+import { BaseParams } from './types/paramTypes.js';
 
 export interface Base {
   connect(): void;
-  sign(params: SignParams): Promise<any>;
+  sign(params: any): Promise<any>;
   getPublicKey(params: BaseParams): Promise<any>;
   getAddress(params: BaseParams): Promise<any>;
   getVersion(params?: BaseParams): Promise<any>;
