+++ packages/bitcore-hardware/src/ledger.ts
@@ -1,11 +1,9 @@
 import { createRequire } from 'module';
-import BIP32Factory from 'bip32';
-import { Psbt, payments } from 'bitcoinjs-lib';
+import { Psbt } from 'bitcoinjs-lib';
 import {
   Observable,
   lastValueFrom
 } from 'rxjs';
-import * as ecc from 'tiny-secp256k1';
 import { Base } from './base.js';
 import { dmk } from './dmk.js';
 import { BaseParams } from './types/paramTypes.js';
@@ -20,7 +18,6 @@ const {
   SignerBtcBuilder
 } = require('@ledgerhq/device-signer-kit-bitcoin');
 
-const bip32 = BIP32Factory(ecc);
 
 export default class Ledger implements Base {
   device: any;
@@ -97,33 +94,7 @@ export default class Ledger implements Base {
     return result.output.extendedPublicKey;
   }
 
-  async sign() {
-    const psbt = new Psbt();
-
-    // Derive the specific child key for external chain (0), address index (0)
-    const childNode = bip32.fromBase58(await this.getPublicKey({ index: 0 })).derive(0).derive(0);
-    const pubkey = Buffer.from(childNode.publicKey);
-
-    // Compute the P2WPKH scriptPubKey from the derived child key
-    const p2wpkh = payments.p2wpkh({ pubkey: pubkey });
-    const script = p2wpkh.output!;
-
-    const masterFingerprint = await this.getMasterKeyFingerprint();
-    psbt.addInput({
-      hash: '78519a191327dfdc0c2ea64a04d09d87c3909ce8365d0e0c0dbd0bc80d0405b4',
-      index: 0,
-      witnessUtxo: {
-        script,
-        value: 1000,
-      },
-      bip32Derivation: [{
-        masterFingerprint: Buffer.from(masterFingerprint.buffer, masterFingerprint.byteOffset, masterFingerprint.byteLength),
-        pubkey,
-        path: "m/84'/0'/0'/0/0",
-      }],
-    });
-    psbt.addOutput({ address: 'bc1qj86hpgprdudkks84y52vdenz86kd26stkssrcq', value: 1000 });
-
+  async sign(psbt: Psbt) {
     const ob: Observable<any> = this.signer.signTransaction(
       new DefaultWallet("84'/0'/0'", 'wpkh(@0/**)'),
       psbt
